+++ tests/Unit/AuthTest.php
@@ -6,13 +6,11 @@ namespace Kreait\Firebase\Tests\Unit;
 
 use Firebase\Auth\Token\Domain\Generator;
 use Firebase\Auth\Token\Domain\Verifier;
-use Firebase\Auth\Token\Exception\InvalidToken;
 use Firebase\Auth\Token\Exception\IssuedInTheFuture;
 use Kreait\Firebase\Auth;
 use Kreait\Firebase\Auth\ApiClient;
 use Kreait\Firebase\Tests\UnitTestCase;
 use Lcobucci\JWT\Token;
-use Prophecy\Argument;
 
 /**
  * @internal
@@ -85,36 +83,7 @@ class AuthTest extends UnitTestCase
             ->method('verifyIdToken')
             ->willReturn($token);
 
-        $verifiedToken = $this->auth->verifyIdToken('foo', false, true);
+        $verifiedToken = $this->auth->verifyIdToken('foo', false);
         $this->assertSame($token, $verifiedToken);
-    }
-
-    public function testAllowAuthTimeToBeInTheFuture()
-    {
-        $tokenProphecy = $this->prophesize(Token::class);
-        $tokenProphecy->getClaim('auth_time', Argument::any())->willReturn(\date('U'));
-
-        $token = $tokenProphecy->reveal();
-
-        $this->idTokenVerifier
-            ->expects($this->once())
-            ->method('verifyIdToken')
-            ->willThrowException(new InvalidToken($token, 'authentication time'));
-
-        $verifiedToken = $this->auth->verifyIdToken($token, false, false);
-        $this->assertSame($token, $verifiedToken);
-    }
-
-    public function testDoNotAllowTimeInconsistenciesAndCheckRevokationStatusAtTheSameTime()
-    {
-        $token = $this->prophesize(Token::class);
-
-        $this->idTokenVerifier
-            ->expects($this->once())
-            ->method('verifyIdToken')
-            ->willReturn($token->reveal());
-
-        $this->expectException(InvalidToken::class);
-        $this->auth->verifyIdToken($token->reveal(), true, true);
     }
 }